Moved files and created MKI directory
