+++ x-pack/test_serverless/functional/test_suites/security/config.cloud_security_posture.agentless.ts
@@ -16,6 +16,8 @@ export default createTestConfig({
     reportName: 'Serverless Security Cloud Security Agentless Onboarding Functional Tests',
   },
   kbnServerArgs: [
+    `--xpack.cloud.serverless.project_id=some_fake_project_id`,
+
     `--xpack.fleet.packages.0.name=cloud_security_posture`,
     `--xpack.fleet.packages.0.version=${CLOUD_CREDENTIALS_PACKAGE_VERSION}`,
     `--xpack.fleet.agentless.enabled=true`,
@@ -26,11 +28,10 @@ export default createTestConfig({
     `--xpack.fleet.agentPolicies.0.id=agentless`,
     `--xpack.fleet.agentPolicies.0.name=agentless`,
     `--xpack.fleet.agentPolicies.0.package_policies=[]`,
-    `--xpack.cloud.serverless.project_id=some_fake_project_id`,
     `--xpack.fleet.agentPolicies.0.is_default=true`,
     `--xpack.fleet.agentPolicies.0.is_default_fleet_server=true`,
 
-    // Serverless Agentless API
+    `--xpack.fleet.agentless.enabled=true`,
     `--xpack.fleet.agentless.api.url=http://localhost:8089`,
     `--xpack.fleet.agentless.api.tls.certificate=${KBN_CERT_PATH}`,
     `--xpack.fleet.agentless.api.tls.key=${KBN_KEY_PATH}`,
