+++ x-pack/test_serverless/functional/test_suites/security/ftr/cloud_security_posture/mki_serverless_quality_gates/index.ts
@@ -1,0 +1,17 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { FtrProviderContext } from '../../../../../ftr_provider_context';
+
+export default function ({ loadTestFile }: FtrProviderContext) {
+  describe('cloud_security_posture', function () {
+    this.tags(['cloud_security_posture']);
+
+    // do not resolve files which are ending with `.essentials.ts`
+    loadTestFile(require.resolve('./agentless/mki_create_agent'));
+  });
+}
